rename gossip blob cache to data availability checker
