+++ beacon_node/beacon_chain/src/builder.rs
@@ -4,7 +4,7 @@ use crate::eth1_chain::{CachingEth1Backend, SszEth1};
 use crate::eth1_finalization_cache::Eth1FinalizationCache;
 use crate::fork_choice_signal::ForkChoiceSignalTx;
 use crate::fork_revert::{reset_fork_choice_to_finalization, revert_to_fork_boundary};
-use crate::gossip_blob_cache::DataAvailabilityChecker;
+use crate::data_availability_checker::DataAvailabilityChecker;
 use crate::head_tracker::HeadTracker;
 use crate::migrate::{BackgroundMigrator, MigratorConfig};
 use crate::persisted_beacon_chain::PersistedBeaconChain;
